feat: merge MonthSettlement
