+++ src/main/java/com/timeToast/timeToast/dto/month_settlement/response/MonthSettlementCreatorResponses.java
@@ -1,0 +1,8 @@
+package com.timeToast.timeToast.dto.month_settlement.response;
+
+import java.util.List;
+
+public record MonthSettlementCreatorResponses(
+        List<MonthSettlementCreatorResponse> monthSettlementResponses
+){
+}
