+++ src/main/java/com/timeToast/timeToast/dto/month_settlement/response/MonthSettlementResponses.java
@@ -2,7 +2,8 @@ package com.timeToast.timeToast.dto.month_settlement.response;
 
 import java.util.List;
 
-public record MonthSettlementResponses (
+public record MonthSettlementResponses(
         List<MonthSettlementResponse> monthSettlementResponses
-){
+
+) {
 }
