+++ src/main/java/com/timeToast/timeToast/repository/month_settlement/MonthSettlementJpaRepository.java
@@ -1,4 +1,4 @@
-package com.timeToast.timeToast.repository.monthSettlement;
+package com.timeToast.timeToast.repository.month_settlement;
 
 import com.timeToast.timeToast.domain.month_settlement.MonthSettlement;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,8 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 import java.time.LocalDate;
-import java.time.YearMonth;
-import java.util.List;
 
 public interface MonthSettlementJpaRepository extends JpaRepository<MonthSettlement, Long> {
     List<MonthSettlement> findAllByMemberId(final long memberId);
