--- src/main/java/com/timeToast/timeToast/service/monthSettlement/MonthSettlementService.java
@@ -1,8 +1,0 @@
-package com.timeToast.timeToast.service.monthSettlement;
-
-import com.timeToast.timeToast.dto.month_settlement.request.MonthSettlementRequest;
-import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementResponses;
-
-public interface MonthSettlementService {
-    MonthSettlementResponses getMonthSettlementByYearMonth(final MonthSettlementRequest monthSettlementRequest);
-}
