+++ src/main/java/com/timeToast/timeToast/service/month_settlement/MonthSettlementServiceImpl.java
@@ -1,27 +1,29 @@
 package com.timeToast.timeToast.service.month_settlement;
 
-import com.timeToast.timeToast.domain.icon.icon.Icon;
-import com.timeToast.timeToast.domain.icon.icon_group.IconGroup;
+import com.timeToast.timeToast.domain.enums.monthSettlement.SettlementState;
 import com.timeToast.timeToast.domain.member.member.Member;
 import com.timeToast.timeToast.domain.month_settlement.MonthSettlement;
-import com.timeToast.timeToast.domain.payment.Payment;
-import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupOrderedResponse;
-import com.timeToast.timeToast.dto.icon.icon_group.response.IconGroupOrderedResponses;
-import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementDetailResponse;
+import com.timeToast.timeToast.dto.month_settlement.request.MonthSettlementRequest;
+import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementCreatorResponse;
+import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementCreatorResponses;
 import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementResponse;
 import com.timeToast.timeToast.dto.month_settlement.response.MonthSettlementResponses;
 import com.timeToast.timeToast.repository.icon.icon.IconRepository;
 import com.timeToast.timeToast.repository.icon.icon_group.IconGroupRepository;
 import com.timeToast.timeToast.repository.member.member.MemberRepository;
-import com.timeToast.timeToast.repository.monthSettlement.MonthSettlementRepository;
+import com.timeToast.timeToast.repository.month_settlement.MonthSettlementRepository;
 import com.timeToast.timeToast.repository.payment.PaymentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -36,16 +38,62 @@ public class MonthSettlementServiceImpl implements MonthSettlementService {
 
     @Transactional(readOnly = true)
     @Override
-    public MonthSettlementResponses getMonthSettlements(final long creatorId) {
+    public MonthSettlementCreatorResponses getCreatorMonthSettlements(final long creatorId) {
         List<MonthSettlement> monthSettlements = monthSettlementRepository.findAllByMemberId(creatorId);
-        List<MonthSettlementResponse> monthSettlementResponses = new ArrayList<>();
+        List<MonthSettlementCreatorResponse> monthSettlementResponses = new ArrayList<>();
 
         monthSettlements.forEach(monthSettlement -> {
-            monthSettlementResponses.add(MonthSettlementResponse.from(monthSettlement));
+            monthSettlementResponses.add(MonthSettlementCreatorResponse.from(monthSettlement));
         });
+        return new MonthSettlementCreatorResponses(monthSettlementResponses);
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public MonthSettlementResponses getMonthSettlementByYearMonth(final MonthSettlementRequest monthSettlementRequest) {
+        List<MonthSettlementResponse> monthSettlementResponses = new ArrayList<>();
+        monthSettlementRepository.findAllByYearMonth(LocalDate.of(monthSettlementRequest.year(), monthSettlementRequest.month(),1))
+                .stream().collect(Collectors.toMap(
+                        MonthSettlement::getMemberId,
+                        response -> response,
+                        (existing, replacement) -> existing
+                )).values().stream().toList().forEach(
+                        monthSettlement -> {
+                            Optional<Member> creator = memberRepository.findById(monthSettlement.getMemberId());
+
+                            if(creator.isPresent()) {
+                                monthSettlementResponses.add(MonthSettlementResponse.builder()
+                                        .memberId(monthSettlement.getMemberId())
+                                        .nickname(creator.get().getNickname())
+                                        .profileUrl(creator.get().getMemberProfileUrl())
+                                        .settlementState(SettlementState.BEFORE_SETTLEMENT)
+                                        .build());
+                            }
+                        }
+                );
+
+
+
         return new MonthSettlementResponses(monthSettlementResponses);
     }
 
+
+
+    @Scheduled(cron = "0 0 0 1 * *")
+    @Transactional
+    public void updateMonthSettlement() {
+        paymentRepository.findAllByMonthlyPayments(LocalDate.now().minusMonths(1), LocalDate.now()).forEach(
+                paymentDto -> monthSettlementRepository.save(
+                        MonthSettlement.builder()
+                                .memberId(paymentDto.memberId())
+                                .iconGroupId(paymentDto.itemId())
+                                .yearMonth(LocalDate.now())
+                                .revenue(paymentDto.salesRevenue())
+                                .settlement((long) (paymentDto.salesRevenue()*0.7))
+                                .settlementState(SettlementState.BEFORE_SETTLEMENT)
+                                .build()));
+
+    }
 
 //    @Transactional(readOnly = true)
 //    @Override
